# Update mdbook to 0.4.37