--- a/src/tools/linkchecker/tests/broken_intra_doc_link/foo.html
+++ b/src/tools/linkchecker/tests/broken_intra_doc_link/foo.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+[<code>std::ffi::CString</code>]
+</body>
+</html>