@ehuss 2024-02-09
@rustbot ready

OK, this is ready now. I added some commits to update the linkchecker to use an actual HTML parser instead of a regex. This was necessary because the new pulldown_cmark changed some escaping behaviors, and the regex couldn't handle random `href="` within normal text. There is a performance penalty, going from about 6 seconds to 18 seconds, but should still be tolerable. I'm not sure if html5ever is really the best choice, but it works well enough I think.

Comparison of the behavior:

Before:
checked links in: 6.7s
number of HTML files scanned: 35034
number of HTML redirects found: 10570
number of links checked: 3009436
number of links ignored due to external: 143040
number of links ignored due to exceptions: 10
number of intra doc links ignored: 7
errors found: 0

After:
checked links in: 18.8s
number of HTML files scanned: 35034
number of HTML redirects found: 10570
number of links checked: 3009439
number of links ignored due to external: 143040
number of links ignored due to exceptions: 10
number of intra doc links ignored: 7
errors found: 0