@Mark-Simulacrum 2024-02-11
Performance hit is unfortunate but seems indeed acceptable.

@bors r+